send-email/format-patch: add qemu-devel mailing list and force v1 revision
We only allow v2+/v1.x revision patch to be a follow-up patchset.
However, sometimes, we want to generate v1 patch and send it to some
message thread. So add an option to allow v1 patchset to be a
follow-up patchset.

qemu-devel mailing list is added with ACPI maintainers Cced.

Signed-off-by: Lv Zheng <zetalog@gmail.com>